First attempt at a fix for multithreading-related GC issues on Linux.  This was done by requiring the compiler runtime to define several new functions:
    extern (C) void* os_query_stackBottom();
    extern (C) void* os_query_stackTop();
    extern (C) void[] os_query_staticData()

Some additional logic was added to the Linux code to determine the stackBottom another way within thread_entryPoint, as __libc_stack_end (which may be used by os_query_stackBottom) doesn't appear to always be valid for new threads.

A later change may be to define generic memory mapping functions somewhere as well, as this would allow essentially all special logic to be removed from the GC code and thus eliminate the need for separate DMD and GDC garbage collectors.