Fix for Ticket #68: Regexp "not" causes an assertion failure
The ptr and len members of BitArray have been made public to preserve the original Phobos logic for a particular function in Regex.  This approach was chosen instead of attempting to fix the workaround code because the D ABI actually specifies that arrays should have those two members public, and it seemed a potentially good idea to remain consistent with this despite the data hiding concerns it raises.