moved the streaming text iterators to tango.text.stream, and updated all known references.
Also removed the shortcut alias names for these templates. The alias causes codegen for the specified type, which is fine if you happen to be working with the alias (but not if you're working with wchar or dchar). Updated the codebase for this change also. I wish there were a good way to hide the template syntax for such things; some way that supports a debugger and does not generate a code instance until actually used ...