Finally removed the struct around tango.text.Text. It is now IFTI enabled, and renamed to be tango.text.Goodies instead (to force a manual intervention upon referring modules).
Note that some of the functions have changed from returning -1 on failure to returning 0 instead. This means a success return is now 1-based rather than 0-based. In actual index-usage it makes little difference overall, but it makes the functions easier to use in trivial cases where the result can be used directly in a boolean expression