Merged in final GDC 0.23 changes.  These changes have also been applied to the DMD runtime where applicable.  As most of these changes were replacing instances of int/uint with ptrdiff_t/size_t, there should be no noticeable difference in DMD behavior from these changes.
Please note that some changes were necessary to the GC as well, and I am trusting that Dave Friedman got things right.  Again, most of these were changes from uint to size_t, but things here seem a bit less clear-cut than most of the changes to the GDC runtime itself.

Finally, it is worth being aware that of the GC code, only gcx.d is really pretty much the same between Tango and Phobos/GPhobos.  The remaining files are split among the GC directory (gc/basic), and the files lifetime.d and memory.d in the compiler runtime code.  The Tango design is far cleaner, but the structural differences complicate merging changes in this area.  I've triple-checked all changes for this merge, but if there are any functional problems, I expect them to concern these GC-related files.