Behavior of the runtime has been modified as follows:
The lifetime of a process is much like the lifetime of an object: it is initialized, then used, then destroyed.  If initialization fails, the successive two steps are never reached.  However, if initialization succeeds, then cleanup will occur even if the use step fails in some way.  Here, the use phase consists of running the user's main function.  If main terminates with an exception, the exception is handled and then cleanup begins.  An exception thrown during cleanup, however, will abort the cleanup process.

In essence, user main is now evaluated separately from the module initialization and cleanup code.  So if main exits with an exception, the module dtors will still be run.