Added the following functions: {{{ #!d extern (C) void  rt_attachDisposeEvent( Object obj, void delegate(Object) evt ); extern (C) void  rt_detachDisposeEvent( Object obj, void delegate(Object) evt ); }}} These serve the same basic purpose as Object.notifyRegister() and Object.notifyUnRegister() in Phobos.
Please note that they may only be called for objects which do not have a user-supplied monitor.  The reasoning is that objects with user-defined monitors may have special lifetime or storage requirements that invalidate any reason for the lifetime monitoring feature, so adding the complexity to support this seems pointless.