Closes #1724.
ZlibInput/Output.reset now correctly reassigns the underlying filter source/sink reference.  I also modified ZlibInput.close to not bail if called more than once; this is mainly because read can implicitly close the stream when it reaches the end, and it's a bit unwieldy to expect the programmer to keep track of whether this has or hasn't happened.